backend/requirements.txt
flask
flask-cors
pandas
numpy
prophet
xgboost
scikit-learn
python-dotenv
groq
holidays
openpyxl
matplotlib
seaborn
